Add Changesets (#1)
This PR sets up changesets for this project. To start publishing with
changesets, add your `NPM_TOKEN` as a secret in this repo, and run `pnpm
changeset` the next time you have a change you want to publish. The
changesets docs also mention that you need to make sure [it doesn't
require two-factor
auth](https://github.com/changesets/action#with-publishing) to publish.
That's just a per-project setting in NPM.

We can also configure changesets to publish both packages in unison,
regardless of where the changes happened, if you want the version
numbers to always line up.

I've removed the version from the test and website package.json-s, which
will make the changesets CLI to omit it from lists asking you which
packages you want to publish.

Oh! It would also be nice to add [the changesets github
app](https://github.com/apps/changeset-bot) so you get comments telling
you about the changesets included in a PR!